config: set the cros in the lambda functions

+++ lambda/delete_inventory_item/lambda_function.py
@@ -25,7 +25,23 @@ def lambda_handler(event, context):
 
     try:
         response = table.query(KeyConditionExpression=Key("id").eq(key_value))
+        
+        # Define allowed CORS headers
+        headers = {
+            "Access-Control-Allow-Origin": "*",  # Allow all origins
+            "Access-Control-Allow-Methods": "GET, POST, DELETE, OPTIONS",  # Allowed HTTP methods
+            "Access-Control-Allow-Headers": "Content-Type"  # Allowed request headers
+        }
 
+        # Check if the request method is OPTIONS (CORS pre-flight request)
+        if event['httpMethod'] == 'OPTIONS':
+            # Return 200 response for OPTIONS requests with CORS headers
+            return {
+                'statusCode': 200,
+                'headers': headers,
+                'body': 'CORS pre-flight response'  # Optional body for pre-flight request
+            }
+    
         for item in response.get("Items", []):
             table.delete_item(
                 Key={
@@ -36,6 +52,7 @@ def lambda_handler(event, context):
 
         return {
             "statusCode": 200,
+            'headers': headers,
             "body": json.dumps(f"Item with ID {key_value} deleted successfully."),
         }
